payouts_project/requirements.txt
requests~=2.26.0
Django~=4.2.0
djangorestframework~=3.14.0
drf-spectacular~=0.29.0
python-dotenv~=1.0.1
gunicorn==20.1.0
psycopg2-binary~=2.9.9
celery[redis]~=5.3.4
